requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
billiard==4.2.1
botocore==1.39.4
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-ip-geolocation==1.6.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
frozenlist==1.7.0
gql==3.5.3
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
multidict==6.6.3
packaging==25.0
parameterized==0.9.0
pika==1.3.2
promise==2.3
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pycodestyle==2.5.0
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
wheel==0.45.1
yarl==1.20.1
